fix: serve the EVP origin trial token via a meta tag, mint the nonce server-side
The Origin-Trial header (only sent on /login) meant that a token
registered for the exact origin planer.solvro.pl never activated on
the beta.planer.solvro.pl subdomain we actually test on. Switch to a
meta tag on the root layout with a subdomain-matching token instead.

Also fix the same nonce-timing bug found in the EVP demo project: the
hidden token input was populated via a client-side useEffect fetch,
so it didn't exist in the DOM at initial render - Chrome's EVP
eligibility check apparently only evaluates the form once, at parse
time. Mint the nonce server-side (auth.api.evpGetNonce()) and embed
it directly in the server-rendered HTML instead, behind a Suspense
boundary as Cache Components requires for request-time randomness.
