+++ .env.example
@@ -16,10 +16,6 @@ SMTP_PORT=465
 SMTP_USERNAME=planer@solvro.pl
 SMTP_PASSWORD=
 
-# Chrome Email Verification Protocol origin trial token, issued for the exact
-# origin this app is served from. See https://developer.chrome.com/origintrials
-CHROME_EVP_ORIGIN_TRIAL_TOKEN=
-
 DATABASE_URL=postgresql://planer:planer@localhost:5432/planer
 REDIS_URL=redis://localhost:6379
 BETTER_AUTH_SECRET=secret
