+++ next.config.ts
@@ -11,19 +11,6 @@ const nextConfig: NextConfig = {
     ],
     qualities: [75, 100],
   },
-  headers() {
-    const evpToken = process.env.CHROME_EVP_ORIGIN_TRIAL_TOKEN;
-    if (evpToken === undefined || evpToken.length === 0) {
-      return [];
-    }
-
-    return [
-      {
-        source: "/login",
-        headers: [{ key: "Origin-Trial", value: evpToken }],
-      },
-    ];
-  },
 };
 
 export default nextConfig;
